% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_agents.R
\name{agentEntry}
\alias{agentEntry}
\title{agent entry}
\usage{
agentEntry(agent, i, intervals, positions, ...)
}
\arguments{
\item{agent}{entering agent}

\item{i}{entry count (id)}

\item{intervals}{(data.table) existing market intervals table}

\item{positions}{(data.table) existing market positions table}

\item{...}{additional arguments to agents}
}
\description{
Request agent's next entry position
}
