% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_agents.R
\name{agentSetup}
\alias{agentSetup}
\title{Set up agent for simulation}
\usage{
agentSetup(agent, n, ...)
}
\arguments{
\item{agent}{agent to set up}

\item{n}{number of entries into market}
}
\description{
Set up agent for simulation
}
