% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_agents.R
\name{agentUpdate}
\alias{agentUpdate}
\title{agent update}
\usage{
agentUpdate(agent, update.idx, intervals, ...)
}
\arguments{
\item{agent}{agent to updated}

\item{update.idx}{updated indices in intervals table}

\item{intervals}{(data.table) new market intervals table}

\item{...}{additional arguments to agents}
}
\value{
updated agent
}
\description{
agent update
}
