% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{agglom}
\alias{agglom}
\title{Condenses list of numbers to convenient printable form, e.g. c(1,2,3,5) -> '1-3,5'}
\usage{
agglom(xs)
}
\description{
Condenses list of numbers to convenient printable form, e.g. c(1,2,3,5) -> '1-3,5'
}
