% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{agglom.reduc}
\alias{agglom.reduc}
\title{Reduction function for turning list of integers into lists of consecutive sets}
\usage{
agglom.reduc(xs, x)
}
\description{
Reduction function for turning list of integers into lists of consecutive sets
}
