% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bridcrit.normal}
\alias{bridcrit.normal}
\title{Calculate bridge interval criterion}
\usage{
bridcrit.normal(
  delta,
  xl,
  xr,
  Wl,
  Wr,
  sigma = 1,
  b = 1,
  Mpadj = squash,
  augment = FALSE,
  ...
)
}
\description{
Calculate bridge interval criterion
}
