% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bridcrit.old}
\alias{bridcrit.old}
\title{bridge criterion function (logged)}
\usage{
bridcrit.old(delta, xl, xr, Wl, Wr, sigma = 1, b = 1, ...)
}
\arguments{
\item{delta}{distance from left point}

\item{xl}{position of left point}

\item{xr}{position of right point}

\item{Wl}{value at left point}

\item{Wr}{value at right point}

\item{sigma}{brownian walk *sigma* parameter}

\item{b}{utility function *b* parameter}
}
\description{
General criterion: 0 = -log M' + log(1/2 V') + log( -Ed2u/Edu )
Open criterion: M' = c'(delta)
Open criterion: V' = sigma^2
}
