% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bridcrit}
\alias{bridcrit}
\title{bridge interval criterion function}
\usage{
bridcrit(delta, xl, xr, Wl, Wr, sigma = 1, b = 1, log = TRUE, ...)
}
\arguments{
\item{delta}{distance from left point}

\item{xl}{position of left point}

\item{xr}{position of right point}

\item{Wl}{value at left point}

\item{Wr}{value at right point}

\item{sigma}{brownian walk *sigma* parameter}

\item{b}{utility function *b* parameter}

\item{log}{use logged criterion?}
}
\description{
General criterion is 0 = M' + ( 1/2 * V' * ( -Ed2u/Edu ) )
Logged criterion is 0 = -log M' + log(1/2 V') + log( -Ed2u/Edu )
Bridge criterion: M' = (Wr-Wl)/(xr-xl) + c'(delta) - c'(dbar)
Bridge criterion: V' = (dbar-delta)/(xr-xl) * sigma^2
}
