% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{bridge.Mp.crit}
\alias{bridge.Mp.crit}
\title{Bridge function Mp criterion}
\usage{
bridge.Mp.crit(delta, xl, xr, Wl, Wr, Mpadj = function(...) 0, ...)
}
\arguments{
\item{delta}{distance from left endpoint}

\item{xl}{left endpoint}

\item{xr}{right endpoint}

\item{Wl}{value at left endpoint}

\item{Wr}{value at right endpoint}
}
\description{
critical zero occurs in the Mp part of the bridge criterion
this helps us find it
}
