% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bridjump}
\alias{bridjump}
\title{New value at bridge jump}
\usage{
bridjump(delta, xl, xr, Wl, Wr, sigma = 1, ...)
}
\arguments{
\item{delta}{jump distance from left endpoint}

\item{xl}{left endpoint}

\item{xr}{right endpoint}

\item{Wl}{value of left point}

\item{Wr}{value of right point}

\item{sigma}{brownian standard deviation}
}
\description{
New value at bridge jump
}
