% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize_positions}
\alias{categorize_positions}
\title{Assigns category grade of membership to positions}
\usage{
categorize_positions(x, mix, peak.difference = TRUE)
}
\arguments{
\item{x}{positions}

\item{mix}{category mixture distribution}

\item{peak.difference}{grade of membership = difference from peak GOM within category}
}
\description{
Assigns category grade of membership to positions
}
