% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-game.R
\name{combine_simulations}
\alias{combine_simulations}
\title{Combine list of simulations into a single structure}
\usage{
combine_simulations(sims_list)
}
\arguments{
\item{sims_list}{list of simulations}
}
\description{
Combine list of simulations into a single structure
}
