% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-game.R
\name{create.positions.table}
\alias{create.positions.table}
\title{Create market entries table}
\usage{
create.positions.table(n)
}
\arguments{
\item{n}{number of positions to create}
}
\description{
Fields
 id --- (key) entry sequence
 x --- entry position
 W --- position value
}
