% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_agents.R
\name{create_plans_table}
\alias{create_plans_table}
\title{set up entry plan table}
\usage{
create_plans_table(n)
}
\arguments{
\item{n}{number of positions}
}
\description{
Fields
 id --- entry sequence (corresponds to interval id)
 delta --- best entry in interval
 Eu --- expected utility of entry at delta
}
