% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dct.logged}
\alias{dct.logged}
\title{Log competition function (d/ddelta)}
\usage{
dct.logged(delta)
}
\arguments{
\item{delta}{distance from competitor}
}
\description{
Log competition function (d/ddelta)
}
