% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Ed2u}
\alias{Ed2u}
\title{Expected second derivative of utility function}
\usage{
Ed2u(delta, W0, a = 1, b = 1, ...)
}
\arguments{
\item{a}{a parameter}

\item{b}{b parameter}

\item{m}{wealth}
}
\description{
risk aversion = b * (1 / (1 + a/b exp(bm)))
a controls straightness, b controls bendiness
RA increasing in b, decreasing in a
}
