% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Eubrid}
\alias{Eubrid}
\title{Expected utility on bridge interval}
\usage{
Eubrid(delta, xl, xr, Wl, Wr, sigma = 1, a = 1, b = 1, ...)
}
\arguments{
\item{delta}{distance from left point}

\item{xl}{location of left point}

\item{xr}{location of right point}

\item{Wl}{value of left point}

\item{Wr}{value of right point}

\item{sigma}{brownian walk *sigma* parameter}

\item{a}{utility function *a* parameter}

\item{b}{utility function *b* parameter}
}
\description{
Expected utility on bridge interval
}
