% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{euratio.normal}
\alias{euratio.normal}
\title{-Ed2u / Edu}
\usage{
euratio.normal(exparg, a = 1, b = 1, ...)
}
\arguments{
\item{exparg}{argument inside exponential term}

\item{a}{utility function *a* parameter}

\item{b}{utility function *b* parameter}

\item{...}{capture additional arguments}
}
\description{
-Ed2u / Edu
}
