% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{euratio}
\alias{euratio}
\title{Ratio of -Ed2u / Edu}
\usage{
euratio(exparg, a = 1, b = 1, log = TRUE, ...)
}
\arguments{
\item{exparg}{argument inside exponential term}

\item{a}{utility function *a* parameter}

\item{b}{utility function *b* parameter}

\item{log}{use logarithmic version?}

\item{...}{additional arguments to subfunctions}
}
\description{
Multiplier on variance term, used in criteria calculations.
Logarithmic version relies on log(1 + exp(x)) for positive x.
For large x, this ~= x + exp(-x)
}
