% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-compete-exp.R
\name{exp.compete.competition.functions}
\alias{exp.compete.competition.functions}
\title{Competition function}
\usage{
\method{exp}{compete.competition.functions}()
}
\arguments{
\item{delta}{distance from competitor}
}
\description{
Competition function
}
