% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_definitions.R
\name{expand_positions}
\alias{expand_positions}
\title{expand finals positions to all iterations}
\usage{
expand_positions(positions, sims, iters)
}
\description{
expand finals positions to all iterations
}
