% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{exparg.brid}
\alias{exparg.brid}
\title{Calculate Eu exponential argument on bridge interval}
\usage{
exparg.brid(delta, xl, xr, Wl, Wr, sigma = 1, b = 1, Madj = squash, ...)
}
\description{
Calculate Eu exponential argument on bridge interval
}
