% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{exparg.open}
\alias{exparg.open}
\title{Calculate Eu exponential argument on open interval}
\usage{
exparg.open(delta, W0, sigma = 1, b = 1, Madj = squash, ...)
}
\description{
Calculate Eu exponential argument on open interval
}
