% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{get_goms}
\alias{get_goms}
\title{Get grades of membership for an object according to a mixture}
\usage{
get_goms(x, mix, logp = TRUE)
}
\arguments{
\item{x}{object to categorize}

\item{mix}{mixtures distributions}

\item{logp}{return grades-of-membership in log form}
}
\description{
Get grades of membership for an object according to a mixture
}
