% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{get_goms_salop}
\alias{get_goms_salop}
\title{Get grades of membership for an object according to a mixture (Salop landscape)}
\usage{
get_goms_salop(x, mix, logp = TRUE)
}
\arguments{
\item{x}{object to categorize}

\item{mix}{mixtures distributions}

\item{logp}{return grades-of-membership in log form}
}
\description{
Currently assuming BAMBI
}
