% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{get_mix_parameters}
\alias{get_mix_parameters}
\title{Extract mixture parameters from a mixture object}
\usage{
get_mix_parameters(mix)
}
\description{
Generalizes over the various ways of calculating mixtures that we have
}
