% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{make_categories}
\alias{make_categories}
\title{Make 'optimal' categories for a set of positions}
\usage{
make_categories(
  positions,
  min.iter = positions[agent.id > 1, min(id)],
  max.iter = positions[, max(id)],
  package = "mclust",
  reportf = NULL,
  ...
)
}
\arguments{
\item{positions}{set of market positions}

\item{max.iter}{maximum iteration to categorize}

\item{package}{which clustering package to use}

\item{reportf}{verbosity function (NULL to be quiet)}

\item{...}{capture additional arguments}

\item{init.iter}{initial iteration to categorize (usually after insertions)}
}
\description{
Finds the optimal number and position of categories for the evolution of a market
Optimal in the sense of information criterion
}
