% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{make_categories_runner}
\alias{make_categories_runner}
\title{Runner method for generating categories}
\usage{
make_categories_runner(positions, nsim = NULL, max.iter = NULL, ...)
}
\arguments{
\item{positions}{positions to categorize, as generated by simulation code}

\item{nsim}{number of simulation runs to categorize}

\item{max.iter}{number of iterations per market to categorize}
}
\description{
Runner method for generating categories
}
