% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-compete-exp.R
\name{make_exp_compete_agent}
\alias{make_exp_compete_agent}
\title{Standard utility function agent}
\usage{
make_exp_compete_agent(a = 1, b = 1, sigma = 1, ...)
}
\arguments{
\item{a}{agent utility function a parameter}

\item{b}{agent utility function b parameter}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Agent with utility u(m) = a*m - exp(-b*m)
}
