% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_agents.R
\name{make_gom_agent}
\alias{make_gom_agent}
\title{Category-aware agent}
\usage{
make_gom_agent(a = 1, b = 1, gom.weight = 1, logp = FALSE, ...)
}
\arguments{
\item{a}{agent utility function a parameter}

\item{b}{agent utility function b parameter}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Agent with utility u(m) = a*m - exp(-b*m)
}
