% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-heuristic.R
\name{make_heuristic_agent}
\alias{make_heuristic_agent}
\title{Standard utility function agent}
\usage{
make_heuristic_agent(jump = 1, sigma = 1, ...)
}
\arguments{
\item{jump}{agent jump distance on open intervals}

\item{sigma}{landscape sigma parameter}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Heuristic agent that always enters at midpoint of bridges and at `jump` distance of open intervals
}
