% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-peak.R
\name{make_peak_agent}
\alias{make_peak_agent}
\title{Determiknistic agent}
\usage{
make_peak_agent(peaks = 0, slope = -1, ...)
}
\arguments{
\item{slope}{slope from known peaks (should be negative)}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Deterministic agent that descends down predetermined slopes from known peaks
}
