% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{make_report_function}
\alias{make_report_function}
\title{Factory for passing data to reporting function}
\usage{
make_report_function(f, simi)
}
\description{
Factory for passing data to reporting function
}
