% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-salop.R
\name{make_salop_agent}
\alias{make_salop_agent}
\title{Standard utility function agent on Salop circle model}
\usage{
make_salop_agent(a = 1, b = 1, ...)
}
\arguments{
\item{a}{agent utility function a parameter}

\item{b}{agent utility function b parameter}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Agent with utility u(m) = a*m - exp(-b*m)
}
