% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-softmax.R
\name{make_softmax_agent}
\alias{make_softmax_agent}
\title{Softmax utility agent}
\usage{
make_softmax_agent(a = 1, ...)
}
\arguments{
\item{a}{agent utility function a parameter}

\item{...}{capture additional arguments}

\item{n}{number of entries into the market}
}
\description{
Agent with utility u(m) = log(exp(a*(m/2)) - 1) -- a certain inverse of logistic market share
}
