% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{opencrit.logged}
\alias{opencrit.logged}
\title{Calculate open interval criterion (logged)}
\usage{
opencrit.logged(
  delta,
  W0,
  sigma = 1,
  b = 1,
  Mpadj = squash,
  augment = FALSE,
  ...
)
}
\description{
Calculate open interval criterion (logged)
}
