% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{opencrit.normal}
\alias{opencrit.normal}
\title{Calculate open interval criterion}
\usage{
opencrit.normal(
  delta,
  W0,
  sigma = 1,
  b = 1,
  Mpadj = squash,
  augment = FALSE,
  ...
)
}
\description{
Calculate open interval criterion
}
