% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{opencrit.old}
\alias{opencrit.old}
\title{open criterion function (logged)}
\usage{
opencrit.old(delta, W0, sigma = 1, b = 1, ...)
}
\arguments{
\item{delta}{distance from point}

\item{W0}{value at point}

\item{sigma}{brownian walk *sigma* parameter}

\item{b}{utility function *b* parameter}
}
\description{
General criterion: 0 = -log M' + log(1/2 V') + log( -Ed2u/Edu )
Open criterion: M' = c'(delta)
Open criterion: V' = sigma^2
}
