% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{opencrit}
\alias{opencrit}
\title{open interval criterion function}
\usage{
opencrit(delta, W0, sigma = 1, b = 1, log = TRUE, ...)
}
\arguments{
\item{delta}{distance from point}

\item{W0}{value at point}

\item{sigma}{brownian walk *sigma* parameter}

\item{b}{utility function *b* parameter}

\item{log}{use logged criterion?}
}
\description{
General criterion is 0 = M' + ( 1/2 * V' * ( -Ed2u/Edu ) )
Logged criterion is 0 = -log M' + log(1/2 V') + log( -Ed2u/Edu )
General criterion: 0 = -log M' + log(1/2 V') + log( -Ed2u/Edu )
Open criterion: M' = c'(delta)
Open criterion: V' = sigma^2
}
