% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{openjump}
\alias{openjump}
\title{New value at open jump}
\usage{
openjump(delta, sigma = 1, ...)
}
\arguments{
\item{delta}{jump distance}

\item{sigma}{brownian standard deviation}
}
\description{
New value at open jump
}
