% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_definitions.R
\name{position_statistics}
\alias{position_statistics}
\title{add category definitions to positions table}
\usage{
position_statistics(positions, goms, min.iter = NULL)
}
\description{
add category definitions to positions table
}
