% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{process_reports}
\alias{process_reports}
\title{Reporting function for parallel categorizations}
\usage{
process_reports(f, sims, max.iter)
}
\description{
Reporting function for parallel categorizations
}
