% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-game.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation with specified agents}
\usage{
run_simulation(
  n,
  agents,
  agent.order = rep(seq_along(agents), length.out = n),
  verbose = .verbose$NONE,
  ...
)
}
\arguments{
\item{n}{number of entries into the market (simulation iterations)}

\item{agents}{list of agent types in the market}

\item{agent.order}{agent entry order into the market}

\item{verbose}{verbosity}

\item{...}{additional arguments to agents}
}
\description{
Run simulation with specified agents
}
