% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-game.R
\name{run_simulations}
\alias{run_simulations}
\title{Run multiple simulations}
\usage{
run_simulations(
  nsim,
  n,
  agents = list(make_standard_agent()),
  insert.dt = data.table(x = 0, W = 0),
  seed = 1,
  verbose = .verbose$NONE,
  ...
)
}
\arguments{
\item{nsim}{number of markets to simulation}

\item{n}{number of entries into each market (simulation iterations)}

\item{insert.dt}{list of initial positions to insert}

\item{seed}{control random seed}

\item{verbose}{verbosity}

\item{...}{additional arguments for agent types}

\item{agent.fs}{list of functions to create agent types}
}
\description{
Run multiple simulations
}
