% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.brid}
\alias{search.brid}
\title{Search bridge interval for optima}
\usage{
search.brid(xl, xr, Wl, Wr, eps = 1e-06, verbose = 0, ...)
}
\arguments{
\item{xl}{left endpoint}

\item{xr}{right endpoint}

\item{Wl}{value at left endpoint}

\item{Wr}{value at right endpoint}

\item{eps}{minimum distance from endpoints to search at}

\item{...}{additional arguments to criterion function}

\item{debug}{report debugging information?}
}
\description{
Search bridge interval for optima
}
