% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.brid.regular}
\alias{search.brid.regular}
\title{Regular search procedure for bridge interval.}
\usage{
search.brid.regular(
  f.n,
  f.f,
  upperzero.int,
  loweroptim.int,
  lowerzero.end,
  mid,
  Mp0,
  eps = 1e-06,
  ...
)
}
\arguments{
\item{f.n}{normal criterion function}

\item{f.f}{fallback criterion function}

\item{upperzero.int}{interval containing higher value zero}

\item{loweroptim.int}{interval containing function minimum in lower half}

\item{lowerzero.end}{endpoint of interval containing lower zero}

\item{mid}{midpoint}

\item{Mp0}{crossing point of Mp (derivative of mean term) function component}

\item{eps}{unnecessary argument?}

\item{...}{additional arguments to criterio function}
}
\description{
We look for maxima of the underlying function by looking for zeros of its derivative.
In general, there is either one maximum, or two maxima and one minimum
In latter case, one half of space contains the single higher valued maximum,
while other half of space contains a minimum near the endpoint and a maximum further away.
So we find maximum in upper half, and in the lower half, we look for minimum of criterion function to check if the minima/maxima crossing points exist; if they do, we find them.
}
