% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search.brid.tight}
\alias{search.brid.tight}
\title{Search procedure when Mp0 (mean derivative crossover) is very close to the midpoint}
\usage{
search.brid.tight(f.n, f.f, lo, mid, hi, eps = 1e-06, ...)
}
\arguments{
\item{lo}{left endpoint of search interval}

\item{mid}{midpoint of search interval}

\item{hi}{right endpoint of search interval}

\item{eps}{minimum distance from endpoints to search at}

\item{...}{additional arguments to criterion function (?)}

\item{f}{criterion function}
}
\description{
bridcrit has zeros at midpoint and at mp0, so this causes problems
}
