% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{top_two_categories}
\alias{top_two_categories}
\title{Find GOM for top two categories}
\usage{
top_two_categories(goms)
}
\arguments{
\item{goms}{grades of membership for all categories}
}
\value{
list(goms=GOM for top, second categories; ords=category ids)
}
\description{
Find GOM for top two categories
}
