% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{u}
\alias{u}
\title{Utility function}
\usage{
u(m, a = 1, b = 1)
}
\arguments{
\item{m}{wealth}

\item{a}{a parameter}

\item{b}{b parameter}
}
\description{
risk aversion = b * (1 / (1 + a/b exp(bm)))
a controls straightness, b controls bendiness
RA increasing in b, decreasing in a
}
