% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Euopen}
\alias{Euopen}
\title{Expected utility on open interval}
\usage{
Euopen(delta, W0, sigma = 1, a = 1, b = 1, Madj = squash, ...)
}
\arguments{
\item{delta}{distance from known point}

\item{W0}{value at known point}

\item{sigma}{random walk sigma parameter}

\item{a}{utility function *a* parameter}

\item{b}{utility function *b* parameter}

\item{...}{capture additional arguments}
}
\description{
Expected utility on open interval
}
