% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{euratio.logged}
\alias{euratio.logged}
\title{Logarithm of -Ed2u / Edu}
\usage{
euratio.logged(exparg, a = 1, b = 1, ...)
}
\arguments{
\item{exparg}{argument inside exponential term}

\item{a}{utility function *a* parameter}

\item{b}{utility function *b* parameter}

\item{...}{capture additional arguments}
}
\description{
Relies on log(1 + exp(x)) for positive x. For large x, this ~= x + exp(-x)
}
